[style] package rename
