+++ src/main/java/RacingGame/common/ErrorMessageConstants.java
@@ -1,4 +1,4 @@
-package step3.common;
+package RacingGame.common;
 
 public enum ErrorMessageConstants {
     CONSTRUCTOR_ERROR_MESSAGE("유틸리티 클래스로 인스턴스 생성 불필요");
