+++ src/main/java/RacingGame/common/NumberGenerator.java
@@ -1,4 +1,4 @@
-package step3.common;
+package RacingGame.common;
 
 public interface NumberGenerator {
     int generate();
