+++ src/test/java/RacingGame/CarsManagerTest.java
@@ -1,11 +1,11 @@
-package study.step3;
+package RacingGame;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import step3.model.Car;
-import step3.model.CarsManager;
+import RacingGame.model.Car;
+import RacingGame.model.CarsManager;
 
 import java.util.Arrays;
 import java.util.List;
